refactor: move Awaiter to separate file, minor adjustments to NextFetchEvent
