+++ packages/next/src/server/web/adapter.ts
@@ -415,7 +415,7 @@ export async function adapter(
 
   return {
     response: finalResponse,
-    waitUntil: event[waitUntilSymbol],
+    waitUntil: event[waitUntilSymbol](),
     fetchMetrics: request.fetchMetrics,
   }
 }
